Refactor #unique_teams
This does the exact same thing, it’s just easier to follow (if you
understand how map works!). So, `#map` will return a new array that is
made up of the return value of each of the iterations through your
array. We’re basically making a new array by grabbing each player’s
value for the ‘team’ key. Finally, we call uniq on the array to remove
any duplicates.