Refactor urls for team pages
You don’t need to strip out the spaces from the URLs. They will be
encoded for you automatically. This lets us get rid of the team_title
method and just use the value supplied by the params for the title of
the page. We also just get to delete code! The code with the fewest
bugs is the code that hasn’t been written :)